Added motor temperature and tacho to packet. Fixed a connection issue

+++ vesc_driver/include/vesc_driver/vesc_packet.h
@@ -190,7 +190,7 @@ public:
   double getInputCharge() const;
   double getConsumedPower() const;
   double getInputPower() const;
-  double getPosition() const;
+  uint32_t getPosition() const;
   double getDisplacement() const;
   int getFaultCode() const;
